post

+++ KnjigomatBackend/src/main/java/web/books/controllers/PostController.java
@@ -4,17 +4,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import web.books.exceptions.NotFoundException;
+import web.books.models.dto.Comment;
 import web.books.models.dto.Post;
-import web.books.models.dto.SinglePost;
 import web.books.models.requests.PostRequest;
+import web.books.security.SecurityConsts;
 import web.books.services.PostService;
 import web.books.services.UserService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Page;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/posts")
@@ -41,8 +42,8 @@ public class PostController {
     }
 
     @GetMapping("/{id}")
-    public SinglePost findById(@PathVariable Integer id) throws NotFoundException {
-        return service.findById(id, SinglePost.class);
+    public Post findById(@PathVariable Integer id) throws NotFoundException {
+        return service.findById(id, Post.class);
     }
     @GetMapping("/users/{id}")
     public List<Post> getAllByUserId (@PathVariable Integer id) throws NotFoundException {
@@ -52,26 +53,29 @@ public class PostController {
 
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Integer id) throws NotFoundException {
-        service.delete(id);
+        Post post = service.findById(id, Post.class);
+        if (Objects.equals(post.getUserId(), userService.getCurrentId()) || userService.getCurrentUser().getRole().equals(SecurityConsts.ADMIN)) {
+            service.delete(id);
+        }
     }
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public SinglePost insert(@RequestBody PostRequest record) throws NotFoundException {
-       record.setCreatedTime(new Timestamp(System.currentTimeMillis()));
-        //record.setUserId(userService.getCurrentId());
-       // record.setCondition("new");
-       // record.setUserId(1);
-       // record.setPrice(BigDecimal.valueOf(4.12));
-       // record.setAuthorName("sasa");
-       // record.setCategoryId(1);
-        return service.insert(record, SinglePost.class);
+    public Post insert(@RequestBody PostRequest postRequest) throws NotFoundException {
+       postRequest.setCreatedTime(new Timestamp(System.currentTimeMillis()));
+       postRequest.setUserId(userService.getCurrentId());
+       return service.insert(postRequest, Post.class);
     }
 
     @PutMapping("/{id}")
-    public SinglePost update(@PathVariable Integer id,@RequestBody PostRequest countryRequest) throws NotFoundException{
-        countryRequest.setCreatedTime(new Timestamp(System.currentTimeMillis()));
-        return service.update(id,countryRequest,SinglePost.class);
+    public Post update(@PathVariable Integer id,@RequestBody PostRequest postRequest) throws NotFoundException{
+        postRequest.setCreatedTime(new Timestamp(System.currentTimeMillis()));
+        postRequest.setUserId(userService.getCurrentId());
+        Post post = service.findById(id, Post.class);
+        if (Objects.equals(post.getUserId(), userService.getCurrentId()) || userService.getCurrentUser().getRole().equals(SecurityConsts.ADMIN)) {
+            return service.update(id,postRequest,Post.class);
+        }else {
+            return post;
+        }
     }
-
 
 }
